@@ -4,34 +4,22 @@
   "private": true,
   "homepage": "",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.1.0",
-    "chart.js": "^3.5.1",
-    "jwt-decode": "^3.1.2",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-calendar": "^3.4.0",
-    "react-chartjs-2": "^3.0.4",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.14.2",
-    "react-icons-kit": "^2.0.0",
-    "react-loading-skeleton": "^2.2.0",
-    "react-router-dom": "^5.2.1",
-    "react-scripts": "4.0.3",
-    "react-select": "^4.3.1",
-    "react-time-picker": "^4.4.1",
-    "react-toastify": "^8.0.2",
-    "web-vitals": "^1.1.2",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@mui/icons-material": "^5.10.16",
-    "@mui/material": "^5.10.15",
-    "socket.io-client": "^4.5.4",
-    "react-qr-code": "^2.0.8"
+    "@formatjs/intl-pluralrules": "^5.1.4",
+    "@formatjs/intl-relativetimeformat": "^11.1.4",
+    "@fortawesome/fontawesome-free-webfonts": "^1.0.9",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "autoprefixer": "10.4.5",
+    "bootstrap": "^5.2.2",
+    "mdb-react-ui-kit": "^4.2.0",
+    "mdb-ui-kit": "^5.0.0",
+    "mdbootstrap": "^4.20.0",
+    "node-sass": "^7.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
